[incubator-kie-issues#2065] WIP - Split DMNRuntimeImpl with DMNRuntimeUtils. Extend test cases with clashing complex type
