+++ kie-dmn/kie-dmn-core/src/test/java/org/kie/dmn/core/imports/ImportsTest.java
@@ -21,6 +21,7 @@ package org.kie.dmn.core.imports;
 import java.math.BigDecimal;
 import java.util.Arrays;
 import java.util.Collections;
+import java.util.Map;
 import java.util.function.Function;
 
 import org.junit.jupiter.params.ParameterizedTest;
@@ -46,9 +47,9 @@ public class ImportsTest extends BaseInterpretedVsCompiledTest {
 
     @ParameterizedTest
     @MethodSource("params")
-    void importingModelsWithSameImportAndInputName(boolean useExecModelCompiler) {
+    void importingModelsWithSameImportAndInputNameSimpledData(boolean useExecModelCompiler) {
         init(useExecModelCompiler);
-        String basePath = "valid_models/DMNv1_6/imports/same_import_and_input_name";
+        String basePath = "valid_models/DMNv1_6/imports/same_import_and_input_name_simple_data";
         String dmnImporting = String.format("%s/ImportingModel.dmn", basePath);
         String dmnImported = String.format("%s/ImportedModel.dmn", basePath);
         final DMNRuntime runtime = DMNRuntimeUtil.createRuntimeWithAdditionalResources(dmnImporting,
@@ -61,7 +62,7 @@ public class ImportsTest extends BaseInterpretedVsCompiledTest {
         assertThat(importedModel.hasErrors()).as(DMNRuntimeUtil.formatMessages(importedModel.getMessages())).isFalse();
 
         final DMNModel importingModel = runtime.getModel("https://kie.org/dmn/_054766E3-2098-4E7B-8F48-576B5373F4EE",
-                                                         "DMN_A0738A91-E039-43A3-B892-0A924B41D549");
+                                                         "ImportingModelSimpleType");
         assertThat(importingModel).isNotNull();
         assertThat(importingModel.hasErrors()).as(DMNRuntimeUtil.formatMessages(importingModel.getMessages())).isFalse();
 
@@ -80,6 +81,52 @@ public class ImportsTest extends BaseInterpretedVsCompiledTest {
         context = runtime.newContext();
         context.set("My Input", "this is my input");
         context.set("CLASHING_NAME", 21);
+
+        evaluateAll = runtime.evaluateAll(importingModel, context);
+        assertThat(evaluateAll.hasErrors()).as(DMNRuntimeUtil.formatMessages(evaluateAll.getMessages())).isFalse();
+
+        LOG.debug("{}", evaluateAll);
+        assertThat(evaluateAll.getDecisionResultByName("My Decision")).isNotNull();
+        assertThat(evaluateAll.getDecisionResultByName("My Decision").getResult()).isEqualTo("this is my input: 21");
+    }
+
+    @ParameterizedTest
+    @MethodSource("params")
+    void importingModelsWithSameImportAndInputNameComplexData(boolean useExecModelCompiler) {
+        init(useExecModelCompiler);
+        String basePath = "valid_models/DMNv1_6/imports/same_import_and_input_name_complex_data";
+        String dmnImporting = String.format("%s/ImportingModel.dmn", basePath);
+        String dmnImported = String.format("%s/ImportedModel.dmn", basePath);
+        final DMNRuntime runtime = DMNRuntimeUtil.createRuntimeWithAdditionalResources(dmnImporting,
+                                                                                       this.getClass(),
+                                                                                       dmnImported);
+
+        final DMNModel importedModel = runtime.getModel("https://kie.org/dmn/_EDBCAD2C-5119-462A-8ACA-7E32C18A78EE",
+                                                        "DMN_C66F78A2-4BD7-4BB7-A200-4C61D82AAFBD");
+        assertThat(importedModel).isNotNull();
+        assertThat(importedModel.hasErrors()).as(DMNRuntimeUtil.formatMessages(importedModel.getMessages())).isFalse();
+
+        final DMNModel importingModel = runtime.getModel("https://kie.org/dmn/_054766E3-2098-4E7B-8F48-576B5373F4EE",
+                                                         "ImportingModelComplexType");
+        assertThat(importingModel).isNotNull();
+        assertThat(importingModel.hasErrors()).as(DMNRuntimeUtil.formatMessages(importingModel.getMessages())).isFalse();
+
+        DMNContext context = runtime.newContext(); // old syntax
+        context.set("My Input", "this is my input");
+        Map<String, Object> clashingNameComplexObject =  mapOf(entry("Complex Type Number", 21), entry("Complex Type String", "A string"));
+        context.set("CLASHING_NAME", mapOf(entry("CLASHING_NAME", clashingNameComplexObject)));
+
+        DMNResult evaluateAll = runtime.evaluateAll(importingModel, context);
+        assertThat(evaluateAll.hasErrors()).as(DMNRuntimeUtil.formatMessages(evaluateAll.getMessages())).isFalse();
+
+        LOG.debug("{}", evaluateAll);
+        assertThat(evaluateAll.getDecisionResultByName("My Decision")).isNotNull();
+        assertThat(evaluateAll.getDecisionResultByName("My Decision").getResult()).isEqualTo("this is my input: 21");
+
+
+        context = runtime.newContext();
+        context.set("My Input", "this is my input");
+        context.set("CLASHING_NAME", mapOf(entry("Complex Type Number", 21), entry("Complex Type String", "A string")));
 
         evaluateAll = runtime.evaluateAll(importingModel, context);
         assertThat(evaluateAll.hasErrors()).as(DMNRuntimeUtil.formatMessages(evaluateAll.getMessages())).isFalse();
