+++ kie-dmn/kie-dmn-test-resources/src/test/resources/valid_models/DMNv1_6/imports/same_import_and_input_name_simple_data/ImportedModel.dmn
